requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==1.4.0
    # via
    #   tensorboard
    #   tensorflow
aiohttp==3.8.4
    # via openai
aiosignal==1.3.1
    # via aiohttp
amqp==5.1.1
    # via kombu
argcomplete==1.10.3
    # via textract
astunparse==1.6.3
    # via tensorflow
async-timeout==4.0.2
    # via
    #   aiohttp
    #   redis
attrs==22.2.0
    # via aiohttp
backports-zoneinfo==0.2.1
    # via
    #   pytz-deprecation-shim
    #   tzlocal
beautifulsoup4==4.8.2
    # via textract
billiard==3.6.4.0
    # via celery
cachetools==5.3.0
    # via google-auth
celery==5.2.7
    # via -r requirements.in
certifi==2022.12.7
    # via requests
chardet==3.0.4
    # via
    #   pdfminer-six
    #   textract
charset-normalizer==3.0.1
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   flask
    #   nltk
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
compressed-rtf==1.0.6
    # via extract-msg
docx2txt==0.8
    # via textract
ebcdic==1.1.1
    # via extract-msg
extract-msg==0.28.7
    # via textract
filelock==3.9.0
    # via
    #   huggingface-hub
    #   transformers
flask==2.2.3
    # via -r requirements.in
flatbuffers==23.3.3
    # via tensorflow
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
gast==0.4.0
    # via tensorflow
google-auth==2.16.2
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow
grpcio==1.51.3
    # via
    #   tensorboard
    #   tensorflow
h5py==3.8.0
    # via tensorflow
huggingface-hub==0.12.1
    # via transformers
idna==3.4
    # via
    #   requests
    #   yarl
imapclient==2.1.0
    # via extract-msg
importlib-metadata==6.0.0
    # via
    #   flask
    #   markdown
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
joblib==1.2.0
    # via nltk
keras==2.11.0
    # via tensorflow
kombu==5.2.4
    # via celery
libclang==15.0.6.1
    # via tensorflow
lxml==4.9.2
    # via python-pptx
markdown==3.4.1
    # via tensorboard
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
nltk==3.8.1
    # via -r requirements.in
numpy==1.24.2
    # via
    #   h5py
    #   opt-einsum
    #   tensorboard
    #   tensorflow
    #   transformers
oauthlib==3.2.2
    # via requests-oauthlib
olefile==0.46
    # via extract-msg
openai==0.27.0
    # via -r requirements.in
opt-einsum==3.3.0
    # via tensorflow
packaging==23.0
    # via
    #   huggingface-hub
    #   tensorflow
    #   transformers
pdfminer-six==20191110
    # via textract
pillow==9.4.0
    # via python-pptx
prompt-toolkit==3.0.38
    # via click-repl
protobuf==3.19.6
    # via
    #   tensorboard
    #   tensorflow
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycryptodome==3.17
    # via pdfminer-six
pypdf2==3.0.1
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
python-pptx==0.6.21
    # via textract
pytz==2022.7.1
    # via celery
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via
    #   huggingface-hub
    #   transformers
redis==4.5.1
    # via -r requirements.in
regex==2022.10.31
    # via
    #   nltk
    #   transformers
requests==2.28.2
    # via
    #   huggingface-hub
    #   openai
    #   requests-oauthlib
    #   tensorboard
    #   transformers
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
six==1.12.0
    # via
    #   astunparse
    #   click-repl
    #   google-auth
    #   google-pasta
    #   imapclient
    #   pdfminer-six
    #   tensorflow
    #   textract
sortedcontainers==2.4.0
    # via pdfminer-six
soupsieve==2.4
    # via beautifulsoup4
speechrecognition==3.8.1
    # via textract
tensorboard==2.11.2
    # via tensorflow
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
tensorflow==2.11.0
    # via -r requirements.in
tensorflow-estimator==2.11.0
    # via tensorflow
tensorflow-io-gcs-filesystem==0.31.0
    # via tensorflow
termcolor==2.2.0
    # via tensorflow
textract==1.6.5
    # via -r requirements.in
tokenizers==0.13.2
    # via transformers
tqdm==4.65.0
    # via
    #   huggingface-hub
    #   nltk
    #   openai
    #   transformers
transformers==4.26.1
    # via -r requirements.in
typing-extensions==4.5.0
    # via
    #   huggingface-hub
    #   pypdf2
    #   tensorflow
tzdata==2022.7
    # via pytz-deprecation-shim
tzlocal==4.2
    # via extract-msg
urllib3==1.26.14
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
werkzeug==2.2.3
    # via
    #   flask
    #   tensorboard
wheel==0.38.4
    # via
    #   astunparse
    #   tensorboard
wrapt==1.15.0
    # via tensorflow
xlrd==1.2.0
    # via textract
xlsxwriter==3.0.8
    # via python-pptx
yarl==1.8.2
    # via aiohttp
zipp==3.15.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
